+++ client/src/components/LikeButton/index.ts
@@ -1,0 +1,1 @@
+export { LikeButton } from './LikeButton'
